fixing select2 bug in safari
the select2 widget will generate a dynamic dropdown element and attach it to the end of the body tag. when closing this dropdown tag Safari will jump to the end of the page since that's where the element was initially placed. this fix adds in a special parent element for each field where the dynamically generated dropdown element will live